Move classic theme to its own directory
