+++ src/classic/primer.js
@@ -1,4 +1,4 @@
-const { colors } = require("@primer/primitives");
+const colors = require("./colors.json"); // Based on "@primer/primitives": "2.0.1",
 
 function getColors(style) {
   if (style === "dark") {
